Refactor variation adapter
