--- packages/protvista-variation-adapter/examples/index.html
@@ -1,47 +1,0 @@
-<!doctype html>
-<html class="no-js" lang="">
-
-<head>
-    <meta charset="utf-8">
-    <meta http-equiv="x-ua-compatible" content="ie=edge">
-    <title></title>
-    <meta name="description" content="">
-    <meta name="viewport" content="width=device-width, initial-scale=1">
-
-    <link rel="apple-touch-icon" href="apple-touch-icon.png">
-    <!-- Place favicon.ico in the root directory -->
-</head>
-
-<body>
-    <div>
-        <protvista-variation-adapter>
-            <data-loader>
-                <source src="https://www.ebi.ac.uk/proteins/api/variation/P05067" />
-            </data-loader>    
-        </protvista-variation-adapter>
-    </div>
-
-    <!-- Required for IE11 -->
-    <script src="../../../node_modules/@babel/polyfill/dist/polyfill.min.js"></script>
-    <!-- Web component polyfill (only loads what it needs) -->
-    <script src="../../../node_modules/@webcomponents/webcomponentsjs/webcomponents-lite.js" charset="utf-8"></script>
-    <!-- Required to polyfill modern browsers as code is ES5 for IE... -->
-    <script src="../../../node_modules/@webcomponents/webcomponentsjs/custom-elements-es5-adapter.js" charset="utf-8"></script>
-
-    <script src="../../data-loader/dist/data-loader.js" charset="utf-8" defer></script>
-
-    <!-- TODO: Change for a local-->
-    <script src="https://cdn.jsdelivr.net/npm/protvista-uniprot-entry-adapter/dist/ProtVistaUniProtEntryAdapter.js" charset="utf-8" defer></script>
-
-    <script src="../dist/protvista-variation-adapter.js" charset="utf-8" defer></script>
-
-    <script>
-        document.querySelector('body').addEventListener('load', function(e){
-            console.log('Data loaded');
-            console.log(e.target);
-            console.log(e.detail);
-        });
-    </script>
-</body>
-
-</html>
