+++ packages/protvista-variation-adapter/src/protvista-variation-adapter.js
@@ -1,11 +1,40 @@
 import ProtvistaFeatureAdapter from "protvista-feature-adapter";
-import groupBy from "lodash-es/groupBy";
 import flatten from "lodash-es/flatten";
 import uniqBy from "lodash-es/uniqBy";
 import forOwn from "lodash-es/forOwn";
 import intersectionBy from "lodash-es/intersectionBy";
 
 import filterData, { getFilter, getColor } from "./filters";
+import formatTooltip from "./tooltipGenerators";
+
+const getSourceType = (xrefs, sourceType) => {
+  const xrefNames = xrefs.map(ref => ref.name);
+  if (sourceType === "uniprot" || sourceType === "mixed") {
+    xrefNames.push("uniprot");
+  }
+  return xrefNames;
+};
+
+export const transformData = data => {
+  const { sequence, features } = data;
+  const variants = features.map(variant => ({
+    type: "Variant",
+    accession: variant.genomicLocation,
+    variant: variant.alternativeSequence,
+    start: variant.begin,
+    end: variant.end,
+    tooltipContent: formatTooltip(variant),
+    color: getColor(variant),
+    association: variant.association,
+    sourceType: variant.sourceType,
+    xrefNames: getSourceType(variant.xrefs, variant.sourceType),
+    clinicalSignificances: variant.clinicalSignificances,
+    polyphenScore: variant.polyphenScore,
+    siftScore: variant.siftScore
+  }));
+  if (!variants) return null;
+  return { sequence, variants };
+};
 
 const filterVariants = (filterName, variants) =>
   getFilter(filterName).applyFilter(variants);
@@ -82,148 +111,7 @@ export default class ProtvistaVariationAdapter extends ProtvistaFeatureAdapter {
   }
 
   parseEntry(data) {
-    const { sequence, features } = data;
-    if (!sequence) return;
-
-    const variants = features.map(variant => {
-      return {
-        type: "Variant",
-        accession: variant.genomicLocation,
-        variant: variant.alternativeSequence,
-        start: variant.begin,
-        end: variant.end,
-        tooltipContent: this.formatTooltip(variant),
-        color: getColor(variant),
-        association: variant.association,
-        sourceType: variant.sourceType,
-        xrefNames: ProtvistaVariationAdapter.getSourceType(
-          variant.xrefs,
-          variant.sourceType
-        ),
-        clinicalSignificances: variant.clinicalSignificances,
-        polyphenScore: variant.polyphenScore,
-        siftScore: variant.siftScore
-      };
-    });
-
-    this._adaptedData = { sequence, variants };
-  }
-
-  static getSourceType(xrefs, sourceType) {
-    const xrefNames = xrefs.map(ref => ref.name);
-    if (sourceType === "uniprot" || sourceType === "mixed") {
-      xrefNames.push("uniprot");
-    }
-    return xrefNames;
-  }
-
-  formatTooltip(variant) {
-    if (variant.description)
-      // eslint-disable-next-line no-param-reassign
-      variant.descriptionArray = ProtvistaVariationAdapter.getDescriptionsAsArray(
-        variant.description
-      );
-    return `
-                <h5>Variant</h5><p>${variant.wildType} > ${
-      variant.alternativeSequence
-    }</p>
-                ${
-                  variant.frequency
-                    ? `<h5>Frequency (MAF)</h5><p>${variant.frequency}</p>`
-                    : ``
-                }
-                ${
-                  variant.siftScore
-                    ? `<h5>SIFT</h5><p>${variant.siftPrediction} ${variant.siftScore}</p>`
-                    : ``
-                }
-                ${
-                  variant.polyphenScore
-                    ? `<h5>Polyphen</h5><p>${variant.polyphenPrediction} ${variant.polyphenScore}</p>`
-                    : ``
-                }
-                ${
-                  variant.consequenceType
-                    ? `<h5>Consequence</h5><p>${variant.consequenceType}</p>`
-                    : ``
-                }
-                ${
-                  variant.somaticStatus
-                    ? `<h5>Somatic</h5><p>${
-                        variant.somaticStatus === 0 ? "No" : "Yes"
-                      }</p>`
-                    : ``
-                }
-                ${
-                  variant.genomicLocation
-                    ? `<h5>Location</h5><p>${variant.genomicLocation}</p>`
-                    : ``
-                }
-                ${
-                  variant.sourceType === "UniProt" ||
-                  variant.sourceType === "mixed"
-                    ? `<hr/>${this.getUniProtHTML(variant)}`
-                    : ""
-                }
-                ${
-                  variant.sourceType === "large_scale_study" ||
-                  variant.sourceType === "mixed"
-                    ? `<hr/>${this.getLSSHTML(variant)}`
-                    : ""
-                }
-        `;
-  }
-
-  getUniProtHTML(variant) {
-    return `<h4>UniProt</h4>
-        ${
-          variant.descriptionArray && variant.descriptionArray.UP
-            ? `<h5>Description</h5><p>${variant.descriptionArray.UP.join(
-                "; "
-              )}</p>`
-            : ``
-        }
-        ${variant.ftId ? `<h5>Feature ID</h5><p>${variant.ftId}</p>` : ``}
-        ${
-          variant.association
-            ? this.getDiseaseAssociations(variant.association)
-            : ""
-        }
-        `;
-  }
-
-  getLSSHTML(variant) {
-    return `<h4>Large scale studies</h4>
-        ${
-          variant.descriptionArray && variant.descriptionArray.LSS
-            ? `<h5>Description</h5><p>${variant.descriptionArray.LSS}</p>`
-            : ``
-        }
-        ${
-          variant.frequency
-            ? `<h5>Frequency (MAF)</h5><p>${variant.frequency}</p>`
-            : ``
-        }
-        <h5>Cross-references</h5>${this._basicHelper.formatXrefs(variant.xrefs)}
-        `;
-  }
-
-  getDiseaseAssociations(associations) {
-    return associations
-      .map(
-        association => `
-            <h4>Disease association</h4><p>${association.name}</p>
-            ${
-              association.xrefs
-                ? `<h5>Cross-references</h5>${this._basicHelper.formatXrefs(
-                    association.xrefs
-                  )}`
-                : ""
-            }
-            ${this._basicHelper.getEvidenceFromCodes(association.evidences)}
-        `
-      )
-      .join("");
+    this._adaptedData = transformData(data);
   }
 
   _fireEvent(name, detail) {
@@ -234,24 +122,5 @@ export default class ProtvistaVariationAdapter extends ProtvistaFeatureAdapter {
         cancelable: true
       })
     );
-  }
-
-  // TODO this is horrible. Jie is looking into changing the API so Xrefs
-  // have a description attribute, so we won't have to use concat.
-  static getDescriptionsAsArray(description) {
-    let descriptionArray = description.split(/\[LSS_|\[SWP]: /g);
-    descriptionArray = groupBy(descriptionArray, desc => {
-      if (desc.length === 0) {
-        return "NOTHING";
-      }
-      if (desc.indexOf("]: ") !== -1) {
-        return "LSS";
-      }
-      return "UP";
-    });
-    descriptionArray.LSS = descriptionArray.LSS
-      ? descriptionArray.LSS.join("; ").replace(/]: /g, ": ")
-      : undefined;
-    return descriptionArray;
   }
 }
